Provide special handling of `NAMED` unreserved keyword.
In the context of `NAMED ONLY` the unreserved keyword is interpreted as
a keyword, in all other cases it's a simple identifier. "Named" is a
sufficiently common generic identifier that it's worthwhile to add
special handling for it.

This is a first step in improving how unreserved keywords are handled.